Create multipage app to display basic user guide, closes #5 (#9)
* create multipage app to display basic user guide

* additional notes plus refactoring
