+++ .github/workflows/test.yml
@@ -23,9 +23,8 @@ jobs:
         fetch-depth: 0
     - name: Install dependencies
       run: |
-        python -m pip install --upgrade pip
-        pip install -r requirements.txt  # Adjust this to your requirements file name
+        pip install -r requirements.txt  
     - name: Run tests
       shell: bash -l {0}
       run: |
-        pytest --verbose
+        pytest -vv
